feat: add helper for waiting on task pickup
This adds a helper which is gonna indefinitely wait for a specific task
to be picked up, before continuing. This will be useful for fixing the
failures of tests that are waiting for a valid status of a repo, but can
be broken by other tasks hogging up the task queue.