feat: add custom eslint rule for import restriction
This adds a new custom ESLint rule that will throw linting errors when
any file inside the `_playwright-tests` folder will import any local
code that is outside the PW test folder. This will be helpful for
preventing breakages in our integration tests which pickup tests for
that folder but not the other suff in the repo.