Test: add shared PW test utilities as submodule
This adds a new git submodule under `_playwright-tests/test-utils` that
comes from out BE repo and contains shared test utilities like helpers,
fixtures and the API client. Also changes one test file to use these
utils as POC/showcase and an update of packages was required.
To make working with the submodule easier a new `test-utils.sh` script
was added and documented in the README.