Add basic tags and annotations to custom repo crud
Include playwright-meta-schema, which provides validation of tags and
annotation content.

The library currently supports a few basic metadata annotation sets, and
can be expanded as the team sees more need for metadata.

Tag values will be limited to what is supported by playwright by default
and what is allowed in the new library.

These do not effect test execution, but the data will be available in
reporters.

There is a new command available for yarn to run to do the validation,
and I have included it in github actions before the tests are run.