change: spinner behavior triggered only on data fetching
Spinner's behavior placed in the top right corner of repositories table was triggered
on 4 occasions (isFetching, repositoryParamsLoading, isIntrospecting, isDeletingItems).
Plus even when there was no action, spinner animation kept running.
Except for isFetching, the other actions have different UX feedback tied to them and
triggering the spinner was redundant for them.
This is now changed that spinner appears only when there is
content list query data fetching going on, or content list query is polling data.
It appears mainly on a slow connection when a user navigates between filters of the table.