fix: display delete template modal over template detail page
Previously when on a template details page, clicking the delete action redirected
to the templates list page where it was overlayed with the delete template modal.
This might have confused users, especially when cancelling the action.
Now the delete modal is showed over the template details and only after
confirmation a user is navigated to the templates list page.
And as expected, when the delete action is triggered from the templates list page,
the delete modal is overlayed over the templates list.