fix: stale template request UUIDs
Previously, when a user was creating a template and went
through the whole process to the end, but then returned
and changed architecture or OS version, the old repos' UUIDs
remained in template request state. The issue was in stale useEffect
due to its dependency array not having been set correctly.
Now this has been fixed together with clearing custom repositories
state, when architecture or OS version change.