Do not manage Dropdown state twice
A lot of tests fails that they cannot "see" kebab menu items, even
though they are obviously there. On these we see "aria-hidden=true".

On manual testing it can randomly became true or false when you
open/close it. My suspicion is that we have race with the lines I
removed and "onClick" in "Toggle".